expand pwm max value to 16 bit for avrs

+++ src/avr/gpio.h
@@ -26,8 +26,8 @@ struct gpio_pwm {
     void *reg;
     uint8_t size8;
 };
-struct gpio_pwm gpio_pwm_setup(uint8_t pin, uint32_t cycle_time, uint8_t val);
-void gpio_pwm_write(struct gpio_pwm g, uint8_t val);
+struct gpio_pwm gpio_pwm_setup(uint8_t pin, uint32_t cycle_time, uint16_t val);
+void gpio_pwm_write(struct gpio_pwm g, uint16_t val);
 
 struct gpio_adc {
     uint8_t chan;
